% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoFill.R
\name{AutoFill}
\alias{AutoFill}
\title{Export content as TeX macro.}
\usage{
AutoFill(commandname, content, outfile, mode = "math", appendmode = FALSE)
}
\arguments{
\item{commandname}{Command name.}

\item{content}{Content you would like to export.}

\item{outfile}{File in which you would like to store the macro.}

\item{mode}{If specified as "text", will appear as normal text even in math mode. If this is undesirable, use "math".}

\item{appendmode}{If TRUE, appends to an existing file.}
}
\description{
Export content as TeX macro.
}
