% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadGlobals.R
\name{LoadGlobals}
\alias{LoadGlobals}
\title{Load scalars from a text file into global environment.}
\usage{
LoadGlobals(input, convert = TRUE, ...)
}
\arguments{
\item{input}{path to \code{.txt} file.}

\item{convert}{whether to guess data type. Defaults to \code{TRUE}. If \code{FALSE}, stores values as strings.}

\item{...}{additional arguments to \code{utils::type.convert}.}
}
\description{
Load scalars from a text file into global environment.
}
\details{
Each line of the input file is assumed to have the name of the global variable followed by
a space followed by the value of the global variable. Lines beginning with # are treated as comments.
Lines in the file that do not have at least two words (defined as blocks of text separated by a
space) are ignored. Strings enclosed in double quotes count as one word.
}
