% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTableTxt.R
\name{MakeTableTxt}
\alias{MakeTableTxt}
\title{Load scalars from a text file into global environment.}
\usage{
MakeTableTxt(df, table_name, outdir)
}
\arguments{
\item{df}{the object to be written, preferably a matrix or data frame. If not, it is attempted to coerce df to a data frame.}

\item{table_name}{the name used both as the table tag and as the exported file’s name}

\item{outdir}{the directory path where the table will be saved}
}
\description{
Load scalars from a text file into global environment.
}
\details{
The function first writes a tag line in the form \verb{<tab:table_name>} to the output file, overwriting if necessary.
The table contents are then written on a new line after this tag. The file is tab-delimited and excludes
row and column names. All values are unquoted.
}
