% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTableTxt.R
\name{MakeTableTxt}
\alias{MakeTableTxt}
\title{Load scalars from a text file into global environment.}
\usage{
MakeTableTxt(df, table_name, outdir)
}
\arguments{
\item{df}{the object to be written, preferably a matrix or data frame. If not, it is attempted to coerce df to a data frame.}

\item{table_name}{}

\item{outdir}{}
}
\description{
Load scalars from a text file into global environment.
}
\details{
Each line of the input file is assumed to have the name of the global variable followed by
a space followed by the value of the global variable. Lines beginning with # are treated as comments.
Lines in the file that do not have at least two words (defined as blocks of text separated by a
space) are ignored. Strings enclosed in double quotes count as one word.
}
